+++ drawing/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
         android:label="FTC Simulator"
         android:theme="@style/Theme.AppCompat.DayNight.NoActionBar">
 
-        <activity android:name=".DrawingRunner"
+        <activity android:name=".DrawingPoints"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
